[feat] Add kafka
Server now has the option of publishing events to kafka for
consumption by an event-persister for writing to the db.
This enables single threaded writing and stronger guarantees
around ordering. Will also make it harder to drop events.

Since event writing is moving from strongly consistent to
eventually consistent many of the spec tests were updated
to poll for 5s before failing to wait for propagation. This pattern
will continue as we add read replicas to optimize the read path.